requirements.txt
openai
streamlit
requests
pillow